+++ frontend/src/pages/my-projects/components/create-project-form/libs/types/project-category-to-readable.type.ts
@@ -1,0 +1,12 @@
+import { ProjectCategory } from '~/packages/projects/projects.js';
+
+const projectCategoryToReadable = {
+  [ProjectCategory.ECOMMERICAL]: 'eCommercial',
+  [ProjectCategory.BUSINESS]: 'Business',
+  [ProjectCategory.BLOG]: 'Blog',
+  [ProjectCategory.PORTFOLIO]: 'Portfolio',
+  [ProjectCategory.PERSONAL]: 'Personal',
+  [ProjectCategory.NONPROFIT]: 'Nonprofit',
+};
+
+export { projectCategoryToReadable };
