+++ frontend/src/pages/my-projects/components/create-project-form/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { projectCategoryToReadable } from './project-category-to-readable.type.js';
